Update picolrn.c
adding comment for generate_binary_test.
randomly generate two points/pixels. the point is in the radius of 128 from the origin.
the distance of the two points is smaller than 64.
The training sample is transformed to dim[-128-127]*[-128-127]
The generated two pixels are used to become a binary comparison feature, so there will be many binary features.
The training sets/samples are estimated the WMSE by each of the comparison feature with the two pixel comparison.
After each of the features is evaluated for its WMSE, a minimum WMSE can be found.